Making a couple optimizations to reduce runtime and allocations/gc.

List of changes:
- 54f0807 new function for adding the cross product tensor, that don't constructs the matrix
- 259e75e in `project2plane()` don't collect, just iterate
- b6a18fe use `zip()` in `compatible...()`, which avoids unnecessary allocations
- 25d5067 `@view` where necessary and don't where not

<details>
<summary>Before PR, on master</summary>
<br>

```julia

julia> using RANSAC, RANSACBenchmark, BenchmarkTools
[ Info: Precompiling RANSAC [d6bc97e0-a563-11e9-1861-c573f65f3bc3]
[ Info: Precompiling RANSACBenchmark [37ac5fee-0dc1-428d-87ed-e8867d5d5384]

julia> bc2 = benchmarkcloud2();

julia> pc = PointCloud(bc2.vertices, bc2.normals, 4);

julia> p = ransacparameters(iteration=(τ=100, itermax=20_000, prob_det=0.99,));

julia> extr, _ = ransac(pc, p, true; reset_rand=true);

julia> extr
4-element Array{ShapeCandidate,1}:
 Cand: (sphere, R: 5.0), 2567 ps
 Cand: (cone, ω: 0.7853981633974495), 868 ps
 Cand: (cylinder, R: 1.7899999999999996), 629 ps
 Cand: (plane), 312 ps

julia> @benchmark ransac($pc, $p, true; reset_rand=true) seconds=60
BenchmarkTools.Trial:
  memory estimate:  680.32 MiB
  allocs estimate:  6664208
  --------------
  minimum time:     1.237 s (5.35% GC)
  median time:      1.367 s (5.32% GC)
  mean time:        1.417 s (6.90% GC)
  maximum time:     1.856 s (15.79% GC)
  --------------
  samples:          43
  evals/sample:     1

julia> versioninfo()
Julia Version 1.4.0
Commit b8e9a9ecc6 (2020-03-21 16:36 UTC)
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: Intel(R) Core(TM) i5-3570 CPU @ 3.40GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-8.0.1 (ORCJIT, ivybridge)
```

</details>