@cserteGT3 2020-05-13
That's all the idea I have for now

<details>
<summary>"final" result</summary>
<br>

```julia
julia> using RANSAC, RANSACBenchmark, BenchmarkTools
[ Info: Precompiling RANSAC [d6bc97e0-a563-11e9-1861-c573f65f3bc3]
[ Info: Precompiling RANSACBenchmark [37ac5fee-0dc1-428d-87ed-e8867d5d5384]

julia> bc2 = benchmarkcloud2();

julia> pc = PointCloud(bc2.vertices, bc2.normals, 4);

julia> p = ransacparameters(iteration=(τ=100, itermax=20_000, prob_det=0.99,));

julia> extr, _ = ransac(pc, p, true; reset_rand=true);

julia> extr
4-element Array{ShapeCandidate,1}:
 Cand: (sphere, R: 5.0), 2567 ps
 Cand: (cone, ω: 0.7853981633974495), 868 ps
 Cand: (cylinder, R: 1.7899999999999996), 629 ps
 Cand: (plane), 312 ps

julia> @benchmark ransac($pc, $p, true; reset_rand=true) seconds=60
BenchmarkTools.Trial:
  memory estimate:  320.86 MiB
  allocs estimate:  5000687
  --------------
  minimum time:     1.010 s (3.03% GC)
  median time:      1.049 s (3.43% GC)
  mean time:        1.078 s (6.73% GC)
  maximum time:     1.215 s (15.86% GC)
  --------------
  samples:          56
  evals/sample:     1

julia> versioninfo()
Julia Version 1.4.0
Commit b8e9a9ecc6 (2020-03-21 16:36 UTC)
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: Intel(R) Core(TM) i5-3570 CPU @ 3.40GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-8.0.1 (ORCJIT, ivybridge)
```

</details>